link to netcdf-cxx4 statically

+++ synxflow/lib/include/primitives/Scalar.h
@@ -18,7 +18,7 @@
 
 namespace GC{
 
-typedef float Scalar;
+typedef double Scalar;
 
 }
